chore(suite): move LayoutContext to support as it is not a component
