+++ packages/suite/src/components/suite/Preloader/SuiteLayout/SuiteLayout.tsx
@@ -14,9 +14,9 @@ import { DiscoveryProgress } from 'src/components/wallet';
 import { NavigationBar } from './NavigationBar/NavigationBar';
 import { useLayoutSize, useSelector, useDevice } from 'src/hooks/suite';
 import { useGuide } from 'src/hooks/guide';
+import { LayoutContext, LayoutContextPayload } from 'src/support/suite/LayoutContext';
 import { ModalContextProvider } from 'src/support/suite/ModalContext';
 import { ModalSwitcher } from '../../modals/ModalSwitcher/ModalSwitcher';
-import { LayoutContext, LayoutContextPayload } from './LayoutContext';
 import { useResetScroll } from './useResetScroll';
 import { useAnchorRemoving } from './useAnchorRemoving';
 import { MenuSecondary } from './MenuSecondary';
