+++ packages/suite/src/hooks/suite/useLayout.tsx
@@ -1,8 +1,6 @@
 import { useEffect, useContext, ComponentType } from 'react';
-import {
-    LayoutContext,
-    SideMenuProps,
-} from 'src/components/suite/Preloader/SuiteLayout/LayoutContext';
+
+import { LayoutContext, SideMenuProps } from 'src/support/suite/LayoutContext';
 
 export const useLayout = (
     title?: string,
